+++ examples/simple_react/amplify/auth.ts
@@ -1,7 +1,0 @@
-import { Auth } from '@aws-amplify/backend-auth';
-
-export const auth = new Auth({
-  loginWith: {
-    email: true,
-  },
-});