+++ examples/simple_react/amplify/auth/resource.ts
@@ -1,0 +1,7 @@
+import { defineAuth } from '@aws-amplify/backend-auth';
+
+export const auth = defineAuth({
+  loginWith: {
+    email: true,
+  },
+});