+++ packages/client-config/src/client-config-contributor/graphql_client_config_contributor.ts
@@ -4,17 +4,28 @@ import {
   graphqlOutputKey,
 } from '@aws-amplify/backend-output-schemas';
 import { GraphqlClientConfig } from '../client-config-types/graphql_client_config.js';
+import { ModelIntrospectionSchemaAdapter } from './model_introspection_schema_adapater.js';
 
 /**
  * Translator for the Graphql API portion of ClientConfig
  */
 export class GraphqlClientConfigContributor implements ClientConfigContributor {
   /**
+   * Constructor
+   * @param modelIntrospectionSchemaAdapter the adapter to provide the model introspection schema from s3 uri
+   */
+  constructor(
+    private readonly modelIntrospectionSchemaAdapter: ModelIntrospectionSchemaAdapter
+  ) {}
+
+  /**
    * Given some BackendOutput, contribute the Graphql API portion of the client config
    */
-  contribute = ({
+  contribute = async ({
     [graphqlOutputKey]: graphqlOutput,
-  }: UnifiedBackendOutput): GraphqlClientConfig | Record<string, never> => {
+  }: UnifiedBackendOutput): Promise<
+    GraphqlClientConfig | Record<string, never>
+  > => {
     if (graphqlOutput === undefined) {
       return {};
     }
@@ -29,6 +40,15 @@ export class GraphqlClientConfigContributor implements ClientConfigContributor {
       aws_appsync_conflictResolutionMode:
         graphqlOutput.payload.awsAppsyncConflictResolutionMode,
     };
+
+    const modelIntrospection =
+      await this.modelIntrospectionSchemaAdapter.getModelIntrospectionSchemaFromS3Uri(
+        graphqlOutput.payload.amplifyApiModelSchemaS3Uri
+      );
+
+    if (modelIntrospection) {
+      config.modelIntrospection = modelIntrospection;
+    }
 
     return config;
   };