+++ packages/deployed-backend-client/src/backend_output_client.ts
@@ -1,7 +1,6 @@
 import { unifiedBackendOutputSchema } from '@aws-amplify/backend-output-schemas';
 import { AmplifyClient } from '@aws-sdk/client-amplify';
 import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
-import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
 import { BackendOutputFetcherFactory } from './backend_output_fetcher_factory.js';
 import { BackendIdentifier } from './index.js';
 import { BackendOutputClient } from './backend_output_client_factory.js';
@@ -13,11 +12,14 @@ export class DefaultBackendOutputClient implements BackendOutputClient {
   /**
    * Instantiates a BackendOutputClient
    */
-  constructor(private credentials: AwsCredentialIdentityProvider) {}
+  constructor(
+    private cloudFormationClient: CloudFormationClient,
+    private amplifyClient: AmplifyClient
+  ) {}
   getOutput = async (backendIdentifier: BackendIdentifier) => {
     const outputFetcher = new BackendOutputFetcherFactory(
-      new CloudFormationClient({ credentials: this.credentials }),
-      new AmplifyClient({ credentials: this.credentials })
+      this.cloudFormationClient,
+      this.amplifyClient
     ).getStrategy(backendIdentifier);
     const output = await outputFetcher.fetchBackendOutput();
     return unifiedBackendOutputSchema.parse(output);