+++ packages/deployed-backend-client/src/deployed_backend_client.test.ts
@@ -16,17 +16,15 @@ import {
   stackOutputKey,
   storageOutputKey,
 } from '@aws-amplify/backend-output-schemas';
-import { BackendOutput } from '@aws-amplify/plugin-types';
-import { BackendOutputClientFactory } from '@aws-amplify/deployed-backend-client';
 import {
   BackendDeploymentType,
   BranchBackendIdentifier,
   SandboxBackendIdentifier,
 } from '@aws-amplify/platform-core';
-import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
 import { DefaultBackendOutputClient } from './backend_output_client.js';
 import { DefaultDeployedBackendClient } from './deployed_backend_client.js';
 import { StackIdentifier } from './index.js';
+import { AmplifyClient } from '@aws-sdk/client-amplify';
 import { GetObjectCommand, S3 } from '@aws-sdk/client-s3';
 
 const listStacksMock = {
@@ -34,6 +32,10 @@ const listStacksMock = {
   StackSummaries: [
     {
       StackName: 'amplify-test-testBranch',
+      StackStatus: StackStatus.CREATE_COMPLETE,
+    },
+    {
+      StackName: 'amplify-error-testBranch',
       StackStatus: StackStatus.CREATE_COMPLETE,
     },
     {
@@ -51,7 +53,7 @@ const listStacksMock = {
       ParentId: 'testStackId',
     },
     {
-      StackName: 'amplify-test-testBranch-api',
+      StackName: 'amplify-test-testBranch-data',
       StackStatus: StackStatus.CREATE_FAILED,
       ParentId: 'testStackId',
     },
@@ -69,7 +71,7 @@ const listStackResourcesMock = {
     },
     {
       PhysicalResourceId:
-        'arn:aws:cloudformation:us-east-1:123:stack/amplify-test-testBranch-api/randomString',
+        'arn:aws:cloudformation:us-east-1:123:stack/amplify-test-testBranch-data/randomString',
       ResourceType: 'AWS::CloudFormation::Stack',
     },
     {
@@ -102,7 +104,8 @@ const getOutputMockResponse = {
   [graphqlOutputKey]: {
     payload: {
       awsAppsyncApiEndpoint: 'testAwsAppsyncApiEndpoint',
-      amplifyApiModelSchemaS3Uri: 's3://bucketName/filePath',
+      amplifyApiModelSchemaS3Uri: 's3://bucketName/model-schema.graphql',
+      awsAppsyncApiId: 'awsAppsyncApiId',
     },
   },
 };
@@ -126,13 +129,22 @@ const expectedMetadata = {
     status: BackendDeploymentStatus.FAILED,
     defaultAuthType: undefined,
     additionalAuthTypes: [],
-    graphqlSchema: 's3://bucketName/filePath schema contents!',
+    graphqlSchema: 's3://bucketName/model-schema.graphql schema contents!',
     conflictResolutionMode: undefined,
+    apiId: 'awsAppsyncApiId',
+    modelIntrospectionSchema:
+      's3://bucketName/model-introspection-schema.json schema contents!',
   },
 };
 
 void describe('Deployed Backend Client', () => {
-  const getOutputMock = mock.fn();
+  const mockCfnClient = new CloudFormation();
+  const mockS3Client = new S3();
+  const mockBackendOutputClient = new DefaultBackendOutputClient(
+    mockCfnClient,
+    new AmplifyClient()
+  );
+  const getOutputMock = mock.method(mockBackendOutputClient, 'getOutput');
   let deployedBackendClient: DefaultDeployedBackendClient;
   const cfnClientSendMock = mock.fn();
   const s3ClientSendMock = mock.fn();
@@ -148,20 +160,21 @@ void describe('Deployed Backend Client', () => {
   });
 
   beforeEach(() => {
-    const mockCredentials: AwsCredentialIdentityProvider = async () => ({
-      accessKeyId: 'accessKeyId',
-      secretAccessKey: 'secretAccessKey',
-    });
-    const mockCfnClient = new CloudFormation();
-    const mockS3Client = new S3();
-    getOutputMock.mock.mockImplementation(() => getOutputMockResponse);
+    getOutputMock.mock.mockImplementation(
+      (backendIdentifier: StackIdentifier) => {
+        if (backendIdentifier.stackName === 'amplify-error-testBranch') {
+          throw new Error('Stack template metadata is not a string');
+        }
+        return getOutputMockResponse;
+      }
+    );
     mock.method(mockCfnClient, 'send', cfnClientSendMock);
     mock.method(mockS3Client, 'send', s3ClientSendMock);
 
     getOutputMock.mock.resetCalls();
     cfnClientSendMock.mock.resetCalls();
     s3ClientSendMock.mock.resetCalls();
-    const mockImplementation = (
+    const mockSend = (
       request:
         | ListStacksCommand
         | DescribeStacksCommand
@@ -184,21 +197,12 @@ void describe('Deployed Backend Client', () => {
       throw request;
     };
 
-    cfnClientSendMock.mock.mockImplementation(mockImplementation);
-
-    const backendOutputClientFactoryMock = mock.fn();
-    backendOutputClientFactoryMock.mock.mockImplementation(() => ({
-      getOutput: getOutputMock as unknown as () => Promise<BackendOutput>,
-    }));
-    BackendOutputClientFactory.getInstance =
-      backendOutputClientFactoryMock as unknown as (
-        credentials: AwsCredentialIdentityProvider
-      ) => DefaultBackendOutputClient;
+    cfnClientSendMock.mock.mockImplementation(mockSend);
 
     deployedBackendClient = new DefaultDeployedBackendClient(
-      mockCredentials,
       mockCfnClient,
-      mockS3Client
+      mockS3Client,
+      mockBackendOutputClient
     );
   });
 
@@ -233,7 +237,7 @@ void describe('Deployed Backend Client', () => {
         },
         {
           deploymentType: BackendDeploymentType.SANDBOX,
-          name: 'amplify-test-testBranch-api',
+          name: 'amplify-test-testBranch-data',
           status: BackendDeploymentStatus.FAILED,
           lastUpdated: undefined,
         },
@@ -263,11 +267,16 @@ void describe('Deployed Backend Client', () => {
 });
 
 void describe('Deployed Backend Client pagination', () => {
+  const mockCfnClient = new CloudFormation();
+  const mockS3Client = new S3();
+  const cfnClientSendMock = mock.method(mockCfnClient, 'send');
   let deployedBackendClient: DefaultDeployedBackendClient;
   const listStacksMockFn = mock.fn();
-  const cfnClientSendMock = mock.fn();
-  const s3ClientSendMock = mock.fn();
-  const getOutputMock = mock.fn();
+  const mockBackendOutputClient = new DefaultBackendOutputClient(
+    mockCfnClient,
+    new AmplifyClient()
+  );
+  const getOutputMock = mock.method(mockBackendOutputClient, 'getOutput');
   const returnedSandboxes = [
     {
       deploymentType: BackendDeploymentType.SANDBOX,
@@ -278,12 +287,11 @@ void describe('Deployed Backend Client pagination', () => {
   ];
 
   beforeEach(() => {
-    const mockCredentials: AwsCredentialIdentityProvider = async () => ({
-      accessKeyId: 'accessKeyId',
-      secretAccessKey: 'secretAccessKey',
-    });
     getOutputMock.mock.mockImplementation(
       (backendIdentifier: StackIdentifier) => {
+        if (backendIdentifier.stackName === 'amplify-error-testBranch') {
+          throw new Error('Stack template metadata is not a string');
+        }
         if (backendIdentifier.stackName !== 'amplify-test-sandbox') {
           return {
             ...getOutputMockResponse,
@@ -298,25 +306,13 @@ void describe('Deployed Backend Client pagination', () => {
       }
     );
     getOutputMock.mock.resetCalls();
-    const backendOutputClientFactoryMock = mock.fn();
-    backendOutputClientFactoryMock.mock.mockImplementation(() => ({
-      getOutput: getOutputMock as unknown as () => Promise<BackendOutput>,
-    }));
-    BackendOutputClientFactory.getInstance =
-      backendOutputClientFactoryMock as unknown as (
-        credentials: AwsCredentialIdentityProvider
-      ) => DefaultBackendOutputClient;
 
-    const mockCfnClient = new CloudFormation();
-    const mockS3Client = new S3();
-    mock.method(mockCfnClient, 'send', cfnClientSendMock);
-    mock.method(mockS3Client, 'send', s3ClientSendMock);
     listStacksMockFn.mock.resetCalls();
     listStacksMockFn.mock.mockImplementation(() => {
       return listStacksMock;
     });
     cfnClientSendMock.mock.resetCalls();
-    const mockImplementation = (
+    const mockSend = (
       request: ListStacksCommand | DescribeStacksCommand | DeleteStackCommand
     ) => {
       if (request instanceof ListStacksCommand) {
@@ -334,12 +330,12 @@ void describe('Deployed Backend Client pagination', () => {
       throw request;
     };
 
-    cfnClientSendMock.mock.mockImplementation(mockImplementation);
+    cfnClientSendMock.mock.mockImplementation(mockSend);
 
     deployedBackendClient = new DefaultDeployedBackendClient(
-      mockCredentials,
       mockCfnClient,
-      mockS3Client
+      mockS3Client,
+      mockBackendOutputClient
     );
   });