+++ packages/deployed-backend-client/src/deployed_backend_client.ts
@@ -2,7 +2,6 @@ import {
   BackendOutput,
   UniqueBackendIdentifier,
 } from '@aws-amplify/plugin-types';
-import { AwsCredentialIdentityProvider } from '@aws-sdk/types';
 import {
   ApiAuthType,
   BackendDeploymentStatus,
@@ -17,7 +16,7 @@ import {
   BackendDeploymentType,
   SandboxBackendIdentifier,
 } from '@aws-amplify/platform-core';
-import { BackendOutputClientFactory } from './backend_output_client_factory.js';
+import { BackendOutputClient } from './backend_output_client_factory.js';
 import { getMainStackName } from './get_main_stack_name.js';
 import {
   CloudFormationClient,
@@ -49,9 +48,9 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
    * Constructor for deployment client
    */
   constructor(
-    private readonly credentials: AwsCredentialIdentityProvider,
     private readonly cfnClient: CloudFormationClient,
-    private readonly s3Client: S3Client
+    private readonly s3Client: S3Client,
+    private readonly backendOutputClient: BackendOutputClient
   ) {}
 
   /**
@@ -70,11 +69,13 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
           return stackSummary.StackStatus !== StackStatus.DELETE_COMPLETE;
         })
         .map(async (stackSummary: StackSummary) => {
+          const deploymentType = await this.tryGetDeploymentType(stackSummary);
+
           return {
             name: stackSummary.StackName as string,
             lastUpdated: stackSummary.LastUpdatedTime,
             status: this.translateStackStatus(stackSummary.StackStatus),
-            deploymentType: await this.getDeploymentType(stackSummary),
+            deploymentType,
           };
         });
 
@@ -96,20 +97,22 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     };
   };
 
-  private getDeploymentType = async (
+  private tryGetDeploymentType = async (
     stackSummary: StackSummary
-  ): Promise<BackendDeploymentType> => {
+  ): Promise<BackendDeploymentType | undefined> => {
     const backendIdentifier = {
       stackName: stackSummary.StackName as string,
     };
 
-    const backendOutput: BackendOutput =
-      await BackendOutputClientFactory.getInstance(this.credentials).getOutput(
-        backendIdentifier
-      );
+    try {
+      const backendOutput: BackendOutput =
+        await this.backendOutputClient.getOutput(backendIdentifier);
 
-    return backendOutput[stackOutputKey].payload
-      .deploymentType as BackendDeploymentType;
+      return backendOutput[stackOutputKey].payload
+        .deploymentType as BackendDeploymentType;
+    } catch {
+      return;
+    }
   };
 
   /**
@@ -152,9 +155,7 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     };
 
     const backendOutput: BackendOutput =
-      await BackendOutputClientFactory.getInstance(this.credentials).getOutput(
-        backendIdentifier
-      );
+      await this.backendOutputClient.getOutput(backendIdentifier);
     const stackDescription = await this.cfnClient.send(
       new DescribeStacksCommand({ StackName: stackName })
     );
@@ -199,7 +200,7 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
         nestedStack?.StackName?.includes('storage')
     );
     const apiStack = childStacks.find((nestedStack: StackSummary | undefined) =>
-      nestedStack?.StackName?.includes('api')
+      nestedStack?.StackName?.includes('data')
     );
 
     const backendMetadataObject: BackendMetadata = {
@@ -228,6 +229,18 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     }
 
     if (apiStack) {
+      const schemaFileUri =
+        backendOutput[graphqlOutputKey]?.payload.amplifyApiModelSchemaS3Uri;
+      const modelIntrospectionSchemaUri =
+        backendOutput[graphqlOutputKey]?.payload
+          .amplifyApiModelIntrospectionSchemaS3Uri ||
+        // TODO: Once the introspection schema is in the CFN output, remove the below line and the OR operator above.
+        // GH link: https://github.com/aws-amplify/samsara-cli/issues/395
+        `${schemaFileUri.slice(
+          0,
+          schemaFileUri.lastIndexOf('/')
+        )}/model-introspection-schema.json`;
+
       const additionalAuthTypesString =
         backendOutput[graphqlOutputKey]?.payload
           .awsAppsyncAdditionalAuthenticationTypes;
@@ -244,8 +257,11 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
         additionalAuthTypes,
         conflictResolutionMode: backendOutput[graphqlOutputKey]?.payload
           .awsAppsyncConflictResolutionMode as ConflictResolutionMode,
-        graphqlSchema: await this.fetchGraphqlSchema(
-          backendOutput[graphqlOutputKey]?.payload.amplifyApiModelSchemaS3Uri
+        graphqlSchema: await this.fetchSchema(schemaFileUri),
+        apiId: backendOutput[graphqlOutputKey]?.payload
+          .awsAppsyncApiId as string,
+        modelIntrospectionSchema: await this.fetchSchema(
+          modelIntrospectionSchemaUri
         ),
       };
     }
@@ -253,20 +269,20 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     return backendMetadataObject;
   };
 
-  private fetchGraphqlSchema = async (
-    graphqlSchemaS3Uri: string | undefined
+  private fetchSchema = async (
+    schemaS3Uri: string | undefined
   ): Promise<string> => {
-    if (!graphqlSchemaS3Uri) {
-      throw new Error('graphqlSchemaS3Uri output is not available');
+    if (!schemaS3Uri) {
+      throw new Error('schemaS3Uri output is not available');
     }
 
     // s3://{bucketName}/{fileName}
-    const uriParts = graphqlSchemaS3Uri.split('/');
+    const uriParts = schemaS3Uri.split('/');
     const bucketName = uriParts[2];
     const objectPath = uriParts.slice(3, uriParts.length).join('/');
 
     if (!bucketName || !objectPath) {
-      throw new Error('graphqlSchemaS3Uri is not valid');
+      throw new Error('schemaS3Uri is not valid');
     }
 
     const s3Response = await this.s3Client.send(
@@ -274,9 +290,7 @@ export class DefaultDeployedBackendClient implements DeployedBackendClient {
     );
 
     if (!s3Response.Body) {
-      throw new Error(
-        `s3Response from ${graphqlSchemaS3Uri} does not contain a Body`
-      );
+      throw new Error(`s3Response from ${schemaS3Uri} does not contain a Body`);
     }
 
     return await s3Response.Body?.transformToString();