+++ packages/deployed-backend-client/src/deployed_backend_client_factory.ts
@@ -5,8 +5,12 @@ import {
   BackendDeploymentType,
   SandboxBackendIdentifier,
 } from '@aws-amplify/platform-core';
-import { CloudFormation } from '@aws-sdk/client-cloudformation';
-import { S3 } from '@aws-sdk/client-s3';
+import { CloudFormationClient } from '@aws-sdk/client-cloudformation';
+import {
+  BackendOutputClient,
+  BackendOutputClientFactory,
+} from './backend_output_client_factory.js';
+import { S3Client } from '@aws-sdk/client-s3';
 
 export enum ConflictResolutionMode {
   LAMBDA = 'LAMBDA',
@@ -45,6 +49,8 @@ export type BackendMetadata = {
     defaultAuthType: ApiAuthType;
     additionalAuthTypes: ApiAuthType[];
     conflictResolutionMode?: ConflictResolutionMode;
+    apiId: string;
+    modelIntrospectionSchema: string;
   };
   authConfiguration?: {
     status: BackendDeploymentStatus;
@@ -83,6 +89,20 @@ export type DeployedBackendClient = {
   ) => Promise<BackendMetadata>;
 };
 
+export type DeployedBackendClientOptions = {
+  s3Client: S3Client;
+  cloudFormationClient: CloudFormationClient;
+  backendOutputClient: BackendOutputClient;
+};
+
+export type DeployedBackendCredentialsOptions = {
+  credentials: AwsCredentialIdentityProvider;
+};
+
+export type DeployedBackendClientFactoryOptions =
+  | DeployedBackendCredentialsOptions
+  | DeployedBackendClientOptions;
+
 /**
  * Factory to create a DeploymentClient
  */
@@ -90,11 +110,22 @@ export class DeployedBackendClientFactory {
   /**
    * Returns a single instance of DeploymentClient
    */
-  static getInstance = (
-    credentials: AwsCredentialIdentityProvider
-  ): DeployedBackendClient => {
-    const cfnClient = new CloudFormation(credentials);
-    const s3Client = new S3(credentials);
-    return new DefaultDeployedBackendClient(credentials, cfnClient, s3Client);
-  };
+  static getInstance(
+    options: DeployedBackendClientFactoryOptions
+  ): DeployedBackendClient {
+    if ('backendOutputClient' in options && 'cloudFormationClient' in options) {
+      return new DefaultDeployedBackendClient(
+        options.cloudFormationClient,
+        options.s3Client,
+        options.backendOutputClient
+      );
+    }
+    return new DefaultDeployedBackendClient(
+      new CloudFormationClient(options.credentials),
+      new S3Client(options.credentials),
+      BackendOutputClientFactory.getInstance({
+        credentials: options.credentials,
+      })
+    );
+  }
 }