+++ packages/model-generator/src/generate_api_code.ts
@@ -110,6 +110,10 @@ export const generateApiCode = async (
               types.writeToDirectory(directoryPath),
             ]);
           },
+          getResults: async () => ({
+            ...documents.getResults,
+            ...types.getResults,
+          }),
         };
       }