+++ packages/model-generator/src/index.ts
@@ -1,3 +1,4 @@
 export * from './model_generator.js';
 export * from './create_graphql_document_generator.js';
 export * from './generate_api_code.js';
+export * from './generate_model_introspection_from_schema_uri.js';