+++ packages/backend-auth/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @aws-amplify/backend-auth
 
+## 0.2.0-alpha.9
+
+### Patch Changes
+
+- 59f5ea24: chore: upgrade aws-cdk to 2.100.0
+- Updated dependencies [eac939b5]
+- Updated dependencies [59f5ea24]
+  - @aws-amplify/auth-construct-alpha@0.2.0-alpha.10
+  - @aws-amplify/backend-output-storage@0.1.1-alpha.3
+  - @aws-amplify/plugin-types@0.2.0-alpha.9
+
+## 0.2.0-alpha.8
+
+### Patch Changes
+
+- 7103735b: cdk lib dependency declaration
+- Updated dependencies [7103735b]
+  - @aws-amplify/auth-construct-alpha@0.2.0-alpha.9
+  - @aws-amplify/plugin-types@0.2.0-alpha.8
+
+## 0.2.0-alpha.7
+
+### Minor Changes
+
+- bf24d363: Change Auth external providers'IDs to secret type
+
 ## 0.2.0-alpha.6
 
 ### Patch Changes