+++ packages/backend-deployer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aws-amplify/backend-deployer
 
+## 0.2.0-alpha.7
+
+### Patch Changes
+
+- 59f5ea24: chore: upgrade aws-cdk to 2.100.0
+- Updated dependencies [59f5ea24]
+  - @aws-amplify/plugin-types@0.2.0-alpha.9
+
+## 0.2.0-alpha.6
+
+### Patch Changes
+
+- c78daa11: fix(sandbox): delete should pass the deployment type to deployer
+- Updated dependencies [7103735b]
+  - @aws-amplify/plugin-types@0.2.0-alpha.8
+
 ## 0.2.0-alpha.5
 
 ### Minor Changes