+++ packages/create-amplify/CHANGELOG.md
@@ -1,5 +1,17 @@
 # create-amplify
 
+## 0.2.0-alpha.10
+
+### Patch Changes
+
+- a130ba6a: fix data template
+
+## 0.2.0-alpha.9
+
+### Patch Changes
+
+- c0d7c475: fix create amplify artifact compilation
+
 ## 0.2.0-alpha.8
 
 ### Minor Changes