+++ packages/storage-construct/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/storage-construct-alpha
 
+## 0.1.1-alpha.8
+
+### Patch Changes
+
+- 59f5ea24: chore: upgrade aws-cdk to 2.100.0
+- Updated dependencies [59f5ea24]
+  - @aws-amplify/backend-output-storage@0.1.1-alpha.3
+
+## 0.1.1-alpha.7
+
+### Patch Changes
+
+- 7103735b: cdk lib dependency declaration
+
 ## 0.1.1-alpha.6
 
 ### Patch Changes